requirements.txt
numpy>=1.19.0
scipy>=1.6.0
pyyaml>=5.4.0
matplotlib>=3.3.0
logging 